Cluster ready the output remains

+++ algo/kMeans++/Handling_input_cluster.hpp
@@ -9,6 +9,6 @@
 
 using namespace std;
 
-vector<vector<int>> store_Cluster_data(int, char**);
+vector<vector<double>> store_Cluster_data(int, char**);
 
 #endif
